cli/demo: move the demo cluster code to its own file
Release note: None
